+++ Vector2.h
@@ -6,10 +6,10 @@
 struct Vector2 
 {
     public:
-        int x;
-        int y;
+        double x;
+        double y;
         Vector2 ();
-        Vector2 ( int x, int y );
+        Vector2 ( double x, double y );
         
         Vector2 & Add ( const Vector2 & vec );
         Vector2 & Subtract ( const Vector2 & vec );
@@ -30,7 +30,7 @@ struct Vector2
         Vector2 & operator /= ( const Vector2 & v );
         */
 
-        Vector2 & operator * ( const int & i );
+        Vector2 & operator * ( const double & i );
         Vector2 & Zero();
 
         friend std::ostream & operator << ( std::ostream & ostr, const Vector2 v );
